requirements.txt
Django==4.2
tzdata==2023.3
djangorestframework==3.14.0
selenium==4.9.0
django-crontab==0.7.1
requests==2.30.0
pytest==7.3.1
flake8==6.0.0
black==23.3.0
